Adding a new XMLString implementation that is more memory efficient and offers more ways to deal with the content including more efficient trim operations. It tries to be as compatible as possible with the old XMLString.
